fix: Update SQL procedures and scripts for improved data handling
- Modified the user selection query in datamartUsers.sh to enhance clarity by removing DISTINCT.
- Changed the data type of the qty variable from SMALLINT to INTEGER in the datamartCountries procedure for better compatibility with expected values.

+++ sql/dwh/datamartCountries/datamartCountries_13_createProcedure.sql
@@ -321,7 +321,7 @@ CREATE OR REPLACE PROCEDURE dwh.update_datamart_country (
 LANGUAGE plpgsql
 AS $proc$
  DECLARE
-  qty SMALLINT;
+  qty INTEGER;
   m_todays_activity INTEGER;
   m_last_year_activity CHAR(371);
   m_lastest_open_note_id INTEGER;
